update README
Current README  is prone to tell people using the LFR as the CWD, however, when execute the 'main' program, must use the subdir as the CWD. This is because the code use  relative path '../shader/' to file the shader files. When I tried to build the code, I had faced the problem that 'SHADER ERROR' and puzzled me for a long time. To avoid the similar situations, I recommend adding some detail about how to execute the './main' after building successfully.